feat(api): tool use is GA and available on 3P (#530)
