+++ api.md
@@ -9,6 +9,7 @@ from anthropic.types import (
     ContentBlockStartEvent,
     ContentBlockStopEvent,
     ImageBlockParam,
+    InputJsonDelta,
     Message,
     MessageDeltaEvent,
     MessageDeltaUsage,
@@ -26,6 +27,10 @@ from anthropic.types import (
     TextBlock,
     TextBlockParam,
     TextDelta,
+    Tool,
+    ToolResultBlockParam,
+    ToolUseBlock,
+    ToolUseBlockParam,
     Usage,
 )
 ```
@@ -34,35 +39,3 @@ Methods:
 
 - <code title="post /v1/messages">client.messages.<a href="./src/anthropic/resources/messages.py">create</a>(\*\*<a href="src/anthropic/types/message_create_params.py">params</a>) -> <a href="./src/anthropic/types/message.py">Message</a></code>
 - <code>client.messages.<a href="./src/anthropic/resources/messages.py">stream</a>(\*args) -> MessageStreamManager[MessageStream] | MessageStreamManager[MessageStreamT]</code>
-
-# Beta
-
-## Tools
-
-### Messages
-
-Types:
-
-```python
-from anthropic.types.beta.tools import (
-    InputJsonDelta,
-    RawToolsBetaContentBlockDeltaEvent,
-    RawToolsBetaContentBlockStartEvent,
-    RawToolsBetaMessageStreamEvent,
-    Tool,
-    ToolResultBlockParam,
-    ToolUseBlock,
-    ToolUseBlockParam,
-    ToolsBetaContentBlock,
-    ToolsBetaContentBlockDeltaEvent,
-    ToolsBetaContentBlockStartEvent,
-    ToolsBetaMessage,
-    ToolsBetaMessageParam,
-    ToolsBetaMessageStreamEvent,
-)
-```
-
-Methods:
-
-- <code title="post /v1/messages?beta=tools">client.beta.tools.messages.<a href="./src/anthropic/resources/beta/tools/messages.py">create</a>(\*\*<a href="src/anthropic/types/beta/tools/message_create_params.py">params</a>) -> <a href="./src/anthropic/types/beta/tools/tools_beta_message.py">ToolsBetaMessage</a></code>
-- <code>client.beta.tools.messages.<a href="./src/anthropic/resources/beta/tools/messages.py">stream</a>(\*args) -> ToolsBetaMessageStreamManager[ToolsBetaMessageStream] | ToolsBetaMessageStreamManager[ToolsBetaMessageStreamT]</code>
