--- src/anthropic/resources/beta/__init__.py
@@ -1,33 +1,0 @@
-# File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
-
-from .beta import (
-    Beta,
-    AsyncBeta,
-    BetaWithRawResponse,
-    AsyncBetaWithRawResponse,
-    BetaWithStreamingResponse,
-    AsyncBetaWithStreamingResponse,
-)
-from .tools import (
-    Tools,
-    AsyncTools,
-    ToolsWithRawResponse,
-    AsyncToolsWithRawResponse,
-    ToolsWithStreamingResponse,
-    AsyncToolsWithStreamingResponse,
-)
-
-__all__ = [
-    "Tools",
-    "AsyncTools",
-    "ToolsWithRawResponse",
-    "AsyncToolsWithRawResponse",
-    "ToolsWithStreamingResponse",
-    "AsyncToolsWithStreamingResponse",
-    "Beta",
-    "AsyncBeta",
-    "BetaWithRawResponse",
-    "AsyncBetaWithRawResponse",
-    "BetaWithStreamingResponse",
-    "AsyncBetaWithStreamingResponse",
-]
