--- src/anthropic/resources/beta/tools/__init__.py
@@ -1,33 +1,0 @@
-# File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
-
-from .tools import (
-    Tools,
-    AsyncTools,
-    ToolsWithRawResponse,
-    AsyncToolsWithRawResponse,
-    ToolsWithStreamingResponse,
-    AsyncToolsWithStreamingResponse,
-)
-from .messages import (
-    Messages,
-    AsyncMessages,
-    MessagesWithRawResponse,
-    AsyncMessagesWithRawResponse,
-    MessagesWithStreamingResponse,
-    AsyncMessagesWithStreamingResponse,
-)
-
-__all__ = [
-    "Messages",
-    "AsyncMessages",
-    "MessagesWithRawResponse",
-    "AsyncMessagesWithRawResponse",
-    "MessagesWithStreamingResponse",
-    "AsyncMessagesWithStreamingResponse",
-    "Tools",
-    "AsyncTools",
-    "ToolsWithRawResponse",
-    "AsyncToolsWithRawResponse",
-    "ToolsWithStreamingResponse",
-    "AsyncToolsWithStreamingResponse",
-]
