+++ src/anthropic/types/__init__.py
@@ -7,8 +7,11 @@ from .message import Message as Message
 from .completion import Completion as Completion
 from .text_block import TextBlock as TextBlock
 from .text_delta import TextDelta as TextDelta
+from .tool_param import ToolParam as ToolParam
 from .content_block import ContentBlock as ContentBlock
 from .message_param import MessageParam as MessageParam
+from .tool_use_block import ToolUseBlock as ToolUseBlock
+from .input_json_delta import InputJsonDelta as InputJsonDelta
 from .text_block_param import TextBlockParam as TextBlockParam
 from .image_block_param import ImageBlockParam as ImageBlockParam
 from .message_stop_event import MessageStopEvent as MessageStopEvent
@@ -16,10 +19,12 @@ from .message_delta_event import MessageDeltaEvent as MessageDeltaEvent
 from .message_delta_usage import MessageDeltaUsage as MessageDeltaUsage
 from .message_start_event import MessageStartEvent as MessageStartEvent
 from .message_stream_event import MessageStreamEvent as MessageStreamEvent
+from .tool_use_block_param import ToolUseBlockParam as ToolUseBlockParam
 from .message_create_params import MessageCreateParams as MessageCreateParams
 from .raw_message_stop_event import RawMessageStopEvent as RawMessageStopEvent
 from .raw_message_delta_event import RawMessageDeltaEvent as RawMessageDeltaEvent
 from .raw_message_start_event import RawMessageStartEvent as RawMessageStartEvent
+from .tool_result_block_param import ToolResultBlockParam as ToolResultBlockParam
 from .completion_create_params import CompletionCreateParams as CompletionCreateParams
 from .content_block_stop_event import ContentBlockStopEvent as ContentBlockStopEvent
 from .raw_message_stream_event import RawMessageStreamEvent as RawMessageStreamEvent
