--- src/anthropic/types/beta/__init__.py
@@ -1,3 +1,0 @@
-# File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
-
-from __future__ import annotations
