--- src/anthropic/types/beta/tools/__init__.py
@@ -1,23 +1,0 @@
-# File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
-
-from __future__ import annotations
-
-from .tool_param import ToolParam as ToolParam
-from .tool_use_block import ToolUseBlock as ToolUseBlock
-from .input_json_delta import InputJsonDelta as InputJsonDelta
-from .tools_beta_message import ToolsBetaMessage as ToolsBetaMessage
-from .tool_use_block_param import ToolUseBlockParam as ToolUseBlockParam
-from .message_create_params import MessageCreateParams as MessageCreateParams
-from .tool_result_block_param import ToolResultBlockParam as ToolResultBlockParam
-from .tools_beta_content_block import ToolsBetaContentBlock as ToolsBetaContentBlock
-from .tools_beta_message_param import ToolsBetaMessageParam as ToolsBetaMessageParam
-from .tools_beta_message_stream_event import ToolsBetaMessageStreamEvent as ToolsBetaMessageStreamEvent
-from .raw_tools_beta_message_stream_event import RawToolsBetaMessageStreamEvent as RawToolsBetaMessageStreamEvent
-from .tools_beta_content_block_delta_event import ToolsBetaContentBlockDeltaEvent as ToolsBetaContentBlockDeltaEvent
-from .tools_beta_content_block_start_event import ToolsBetaContentBlockStartEvent as ToolsBetaContentBlockStartEvent
-from .raw_tools_beta_content_block_delta_event import (
-    RawToolsBetaContentBlockDeltaEvent as RawToolsBetaContentBlockDeltaEvent,
-)
-from .raw_tools_beta_content_block_start_event import (
-    RawToolsBetaContentBlockStartEvent as RawToolsBetaContentBlockStartEvent,
-)
