--- src/anthropic/types/beta/tools/message_create_params.py
@@ -1,310 +1,0 @@
-# File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
-
-from __future__ import annotations
-
-from typing import List, Union, Iterable, Optional
-from typing_extensions import Literal, Required, TypedDict
-
-from .tool_param import ToolParam
-from .tools_beta_message_param import ToolsBetaMessageParam
-
-__all__ = [
-    "MessageCreateParamsBase",
-    "Metadata",
-    "ToolChoice",
-    "ToolChoiceToolChoiceAuto",
-    "ToolChoiceToolChoiceAny",
-    "ToolChoiceToolChoiceTool",
-    "MessageCreateParamsNonStreaming",
-    "MessageCreateParamsStreaming",
-]
-
-
-class MessageCreateParamsBase(TypedDict, total=False):
-    max_tokens: Required[int]
-    """The maximum number of tokens to generate before stopping.
-
-    Note that our models may stop _before_ reaching this maximum. This parameter
-    only specifies the absolute maximum number of tokens to generate.
-
-    Different models have different maximum values for this parameter. See
-    [models](https://docs.anthropic.com/en/docs/models-overview) for details.
-    """
-
-    messages: Required[Iterable[ToolsBetaMessageParam]]
-    """Input messages.
-
-    Our models are trained to operate on alternating `user` and `assistant`
-    conversational turns. When creating a new `Message`, you specify the prior
-    conversational turns with the `messages` parameter, and the model then generates
-    the next `Message` in the conversation.
-
-    Each input message must be an object with a `role` and `content`. You can
-    specify a single `user`-role message, or you can include multiple `user` and
-    `assistant` messages. The first message must always use the `user` role.
-
-    If the final message uses the `assistant` role, the response content will
-    continue immediately from the content in that message. This can be used to
-    constrain part of the model's response.
-
-    Example with a single `user` message:
-
-    ```json
-    [{ "role": "user", "content": "Hello, Claude" }]
-    ```
-
-    Example with multiple conversational turns:
-
-    ```json
-    [
-      { "role": "user", "content": "Hello there." },
-      { "role": "assistant", "content": "Hi, I'm Claude. How can I help you?" },
-      { "role": "user", "content": "Can you explain LLMs in plain English?" }
-    ]
-    ```
-
-    Example with a partially-filled response from Claude:
-
-    ```json
-    [
-      {
-        "role": "user",
-        "content": "What's the Greek name for Sun? (A) Sol (B) Helios (C) Sun"
-      },
-      { "role": "assistant", "content": "The best answer is (" }
-    ]
-    ```
-
-    Each input message `content` may be either a single `string` or an array of
-    content blocks, where each block has a specific `type`. Using a `string` for
-    `content` is shorthand for an array of one content block of type `"text"`. The
-    following input messages are equivalent:
-
-    ```json
-    { "role": "user", "content": "Hello, Claude" }
-    ```
-
-    ```json
-    { "role": "user", "content": [{ "type": "text", "text": "Hello, Claude" }] }
-    ```
-
-    Starting with Claude 3 models, you can also send image content blocks:
-
-    ```json
-    {
-      "role": "user",
-      "content": [
-        {
-          "type": "image",
-          "source": {
-            "type": "base64",
-            "media_type": "image/jpeg",
-            "data": "/9j/4AAQSkZJRg..."
-          }
-        },
-        { "type": "text", "text": "What is in this image?" }
-      ]
-    }
-    ```
-
-    We currently support the `base64` source type for images, and the `image/jpeg`,
-    `image/png`, `image/gif`, and `image/webp` media types.
-
-    See [examples](https://docs.anthropic.com/en/api/messages-examples) for more
-    input examples.
-
-    Note that if you want to include a
-    [system prompt](https://docs.anthropic.com/en/docs/system-prompts), you can use
-    the top-level `system` parameter — there is no `"system"` role for input
-    messages in the Messages API.
-    """
-
-    model: Required[str]
-    """The model that will complete your prompt.
-
-    See [models](https://docs.anthropic.com/en/docs/models-overview) for additional
-    details and options.
-    """
-
-    metadata: Metadata
-    """An object describing metadata about the request."""
-
-    stop_sequences: List[str]
-    """Custom text sequences that will cause the model to stop generating.
-
-    Our models will normally stop when they have naturally completed their turn,
-    which will result in a response `stop_reason` of `"end_turn"`.
-
-    If you want the model to stop generating when it encounters custom strings of
-    text, you can use the `stop_sequences` parameter. If the model encounters one of
-    the custom sequences, the response `stop_reason` value will be `"stop_sequence"`
-    and the response `stop_sequence` value will contain the matched stop sequence.
-    """
-
-    system: str
-    """System prompt.
-
-    A system prompt is a way of providing context and instructions to Claude, such
-    as specifying a particular goal or role. See our
-    [guide to system prompts](https://docs.anthropic.com/en/docs/system-prompts).
-    """
-
-    temperature: float
-    """Amount of randomness injected into the response.
-
-    Defaults to `1.0`. Ranges from `0.0` to `1.0`. Use `temperature` closer to `0.0`
-    for analytical / multiple choice, and closer to `1.0` for creative and
-    generative tasks.
-
-    Note that even with `temperature` of `0.0`, the results will not be fully
-    deterministic.
-    """
-
-    tool_choice: ToolChoice
-    """How the model should use the provided tools.
-
-    The model can use a specific tool, any available tool, or decide by itself.
-    """
-
-    tools: Iterable[ToolParam]
-    """[beta] Definitions of tools that the model may use.
-
-    If you include `tools` in your API request, the model may return `tool_use`
-    content blocks that represent the model's use of those tools. You can then run
-    those tools using the tool input generated by the model and then optionally
-    return results back to the model using `tool_result` content blocks.
-
-    Each tool definition includes:
-
-    - `name`: Name of the tool.
-    - `description`: Optional, but strongly-recommended description of the tool.
-    - `input_schema`: [JSON schema](https://json-schema.org/) for the tool `input`
-      shape that the model will produce in `tool_use` output content blocks.
-
-    For example, if you defined `tools` as:
-
-    ```json
-    [
-      {
-        "name": "get_stock_price",
-        "description": "Get the current stock price for a given ticker symbol.",
-        "input_schema": {
-          "type": "object",
-          "properties": {
-            "ticker": {
-              "type": "string",
-              "description": "The stock ticker symbol, e.g. AAPL for Apple Inc."
-            }
-          },
-          "required": ["ticker"]
-        }
-      }
-    ]
-    ```
-
-    And then asked the model "What's the S&P 500 at today?", the model might produce
-    `tool_use` content blocks in the response like this:
-
-    ```json
-    [
-      {
-        "type": "tool_use",
-        "id": "toolu_01D7FLrfh4GYq7yT1ULFeyMV",
-        "name": "get_stock_price",
-        "input": { "ticker": "^GSPC" }
-      }
-    ]
-    ```
-
-    You might then run your `get_stock_price` tool with `{"ticker": "^GSPC"}` as an
-    input, and return the following back to the model in a subsequent `user`
-    message:
-
-    ```json
-    [
-      {
-        "type": "tool_result",
-        "tool_use_id": "toolu_01D7FLrfh4GYq7yT1ULFeyMV",
-        "content": "259.75 USD"
-      }
-    ]
-    ```
-
-    Tools can be used for workflows that include running client-side tools and
-    functions, or more generally whenever you want the model to produce a particular
-    JSON structure of output.
-
-    See our [beta guide](https://docs.anthropic.com/en/docs/tool-use) for more
-    details.
-    """
-
-    top_k: int
-    """Only sample from the top K options for each subsequent token.
-
-    Used to remove "long tail" low probability responses.
-    [Learn more technical details here](https://towardsdatascience.com/how-to-sample-from-language-models-682bceb97277).
-
-    Recommended for advanced use cases only. You usually only need to use
-    `temperature`.
-    """
-
-    top_p: float
-    """Use nucleus sampling.
-
-    In nucleus sampling, we compute the cumulative distribution over all the options
-    for each subsequent token in decreasing probability order and cut it off once it
-    reaches a particular probability specified by `top_p`. You should either alter
-    `temperature` or `top_p`, but not both.
-
-    Recommended for advanced use cases only. You usually only need to use
-    `temperature`.
-    """
-
-
-class Metadata(TypedDict, total=False):
-    user_id: Optional[str]
-    """An external identifier for the user who is associated with the request.
-
-    This should be a uuid, hash value, or other opaque identifier. Anthropic may use
-    this id to help detect abuse. Do not include any identifying information such as
-    name, email address, or phone number.
-    """
-
-
-class ToolChoiceToolChoiceAuto(TypedDict, total=False):
-    type: Required[Literal["auto"]]
-
-
-class ToolChoiceToolChoiceAny(TypedDict, total=False):
-    type: Required[Literal["any"]]
-
-
-class ToolChoiceToolChoiceTool(TypedDict, total=False):
-    name: Required[str]
-    """The name of the tool to use."""
-
-    type: Required[Literal["tool"]]
-
-
-ToolChoice = Union[ToolChoiceToolChoiceAuto, ToolChoiceToolChoiceAny, ToolChoiceToolChoiceTool]
-
-
-class MessageCreateParamsNonStreaming(MessageCreateParamsBase):
-    stream: Literal[False]
-    """Whether to incrementally stream the response using server-sent events.
-
-    See [streaming](https://docs.anthropic.com/en/api/messages-streaming) for
-    details.
-    """
-
-
-class MessageCreateParamsStreaming(MessageCreateParamsBase):
-    stream: Required[Literal[True]]
-    """Whether to incrementally stream the response using server-sent events.
-
-    See [streaming](https://docs.anthropic.com/en/api/messages-streaming) for
-    details.
-    """
-
-
-MessageCreateParams = Union[MessageCreateParamsNonStreaming, MessageCreateParamsStreaming]
