--- src/anthropic/types/beta/tools/raw_tools_beta_content_block_start_event.py
@@ -1,21 +1,0 @@
-# File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
-
-from typing import Union
-from typing_extensions import Literal, Annotated
-
-from ...._utils import PropertyInfo
-from ...._models import BaseModel
-from ...text_block import TextBlock
-from .tool_use_block import ToolUseBlock
-
-__all__ = ["RawToolsBetaContentBlockStartEvent", "ContentBlock"]
-
-ContentBlock = Annotated[Union[TextBlock, ToolUseBlock], PropertyInfo(discriminator="type")]
-
-
-class RawToolsBetaContentBlockStartEvent(BaseModel):
-    content_block: ContentBlock
-
-    index: int
-
-    type: Literal["content_block_start"]
