--- src/anthropic/types/beta/tools/tools_beta_content_block.py
@@ -1,12 +1,0 @@
-# File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
-
-from typing import Union
-from typing_extensions import Annotated
-
-from ...._utils import PropertyInfo
-from ...text_block import TextBlock
-from .tool_use_block import ToolUseBlock
-
-__all__ = ["ToolsBetaContentBlock"]
-
-ToolsBetaContentBlock = Annotated[Union[TextBlock, ToolUseBlock], PropertyInfo(discriminator="type")]
