--- src/anthropic/types/beta/tools/tools_beta_content_block_delta_event.py
@@ -1,10 +1,0 @@
-# File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
-
-
-
-from .raw_tools_beta_content_block_delta_event import RawToolsBetaContentBlockDeltaEvent
-
-__all__ = ["ToolsBetaContentBlockDeltaEvent"]
-
-ToolsBetaContentBlockDeltaEvent = RawToolsBetaContentBlockDeltaEvent
-"""The RawToolsBetaContentBlockDeltaEvent type should be used instead"""
