--- src/anthropic/types/beta/tools/tools_beta_content_block_start_event.py
@@ -1,10 +1,0 @@
-# File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
-
-
-
-from .raw_tools_beta_content_block_start_event import RawToolsBetaContentBlockStartEvent
-
-__all__ = ["ToolsBetaContentBlockStartEvent"]
-
-ToolsBetaContentBlockStartEvent = RawToolsBetaContentBlockStartEvent
-"""The RawToolsBetaContentBlockStartEvent type should be used instead"""
