--- src/anthropic/types/beta/tools/tools_beta_message_stream_event.py
@@ -1,10 +1,0 @@
-# File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
-
-
-
-from .raw_tools_beta_message_stream_event import RawToolsBetaMessageStreamEvent
-
-__all__ = ["ToolsBetaMessageStreamEvent"]
-
-ToolsBetaMessageStreamEvent = RawToolsBetaMessageStreamEvent
-"""The RawToolsBetaMessageStreamEvent type should be used instead"""
