+++ src/anthropic/types/input_json_delta.py
@@ -2,7 +2,7 @@
 
 from typing_extensions import Literal
 
-from ...._models import BaseModel
+from .._models import BaseModel
 
 __all__ = ["InputJsonDelta"]
