+++ src/anthropic/types/message.py
@@ -5,8 +5,7 @@ from typing_extensions import Literal
 
 from .usage import Usage
 from .._models import BaseModel
-from .text_block import TextBlock
-from .content_block import ContentBlock as ContentBlock
+from .content_block import ContentBlock, ContentBlock as ContentBlock
 
 __all__ = ["Message"]
 
@@ -18,11 +17,11 @@ class Message(BaseModel):
     The format and length of IDs may change over time.
     """
 
-    content: List[TextBlock]
+    content: List[ContentBlock]
     """Content generated by the model.
 
     This is an array of content blocks, each of which has a `type` that determines
-    its shape. Currently, the only `type` in responses is `"text"`.
+    its shape.
 
     Example:
 
@@ -62,7 +61,7 @@ class Message(BaseModel):
     This will always be `"assistant"`.
     """
 
-    stop_reason: Optional[Literal["end_turn", "max_tokens", "stop_sequence"]] = None
+    stop_reason: Optional[Literal["end_turn", "max_tokens", "stop_sequence", "tool_use"]] = None
     """The reason that we stopped.
 
     This may be one the following values:
@@ -70,6 +69,7 @@ class Message(BaseModel):
     - `"end_turn"`: the model reached a natural stopping point
     - `"max_tokens"`: we exceeded the requested `max_tokens` or the model's maximum
     - `"stop_sequence"`: one of your provided custom `stop_sequences` was generated
+    - `"tool_use"`: the model invoked one or more tools
 
     In non-streaming mode this value is always non-null. In streaming mode, it is
     null in the `message_start` event and non-null otherwise.
