+++ src/anthropic/types/message_param.py
@@ -8,11 +8,17 @@ from typing_extensions import Literal, Required, TypedDict
 from .content_block import ContentBlock
 from .text_block_param import TextBlockParam
 from .image_block_param import ImageBlockParam
+from .tool_use_block_param import ToolUseBlockParam
+from .tool_result_block_param import ToolResultBlockParam
 
 __all__ = ["MessageParam"]
 
 
 class MessageParam(TypedDict, total=False):
-    content: Required[Union[str, Iterable[Union[TextBlockParam, ImageBlockParam, ContentBlock]]]]
+    content: Required[
+        Union[
+            str, Iterable[Union[TextBlockParam, ImageBlockParam, ToolUseBlockParam, ToolResultBlockParam, ContentBlock]]
+        ]
+    ]
 
     role: Required[Literal["user", "assistant"]]
