+++ src/anthropic/types/raw_content_block_delta_event.py
@@ -1,15 +1,20 @@
 # File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
-from typing_extensions import Literal
+from typing import Union
+from typing_extensions import Literal, Annotated
 
+from .._utils import PropertyInfo
 from .._models import BaseModel
 from .text_delta import TextDelta
+from .input_json_delta import InputJsonDelta
 
-__all__ = ["RawContentBlockDeltaEvent"]
+__all__ = ["RawContentBlockDeltaEvent", "Delta"]
+
+Delta = Annotated[Union[TextDelta, InputJsonDelta], PropertyInfo(discriminator="type")]
 
 
 class RawContentBlockDeltaEvent(BaseModel):
-    delta: TextDelta
+    delta: Delta
 
     index: int
