+++ src/anthropic/types/tool_use_block.py
@@ -2,7 +2,7 @@
 
 from typing_extensions import Literal
 
-from ...._models import BaseModel
+from .._models import BaseModel
 
 __all__ = ["ToolUseBlock"]
