--- tests/api_resources/beta/__init__.py
@@ -1,1 +1,0 @@
-# File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
