+++ tests/api_resources/test_messages.py
@@ -47,6 +47,60 @@ class TestMessages:
             stream=False,
             system="Today's date is 2024-01-01.",
             temperature=1,
+            tool_choice={"type": "auto"},
+            tools=[
+                {
+                    "description": "Get the current weather in a given location",
+                    "name": "x",
+                    "input_schema": {
+                        "type": "object",
+                        "properties": {
+                            "location": {
+                                "description": "The city and state, e.g. San Francisco, CA",
+                                "type": "string",
+                            },
+                            "unit": {
+                                "description": "Unit for the output - one of (celsius, fahrenheit)",
+                                "type": "string",
+                            },
+                        },
+                    },
+                },
+                {
+                    "description": "Get the current weather in a given location",
+                    "name": "x",
+                    "input_schema": {
+                        "type": "object",
+                        "properties": {
+                            "location": {
+                                "description": "The city and state, e.g. San Francisco, CA",
+                                "type": "string",
+                            },
+                            "unit": {
+                                "description": "Unit for the output - one of (celsius, fahrenheit)",
+                                "type": "string",
+                            },
+                        },
+                    },
+                },
+                {
+                    "description": "Get the current weather in a given location",
+                    "name": "x",
+                    "input_schema": {
+                        "type": "object",
+                        "properties": {
+                            "location": {
+                                "description": "The city and state, e.g. San Francisco, CA",
+                                "type": "string",
+                            },
+                            "unit": {
+                                "description": "Unit for the output - one of (celsius, fahrenheit)",
+                                "type": "string",
+                            },
+                        },
+                    },
+                },
+            ],
             top_k=5,
             top_p=0.7,
         )
@@ -121,6 +175,60 @@ class TestMessages:
             stop_sequences=["string", "string", "string"],
             system="Today's date is 2024-01-01.",
             temperature=1,
+            tool_choice={"type": "auto"},
+            tools=[
+                {
+                    "description": "Get the current weather in a given location",
+                    "name": "x",
+                    "input_schema": {
+                        "type": "object",
+                        "properties": {
+                            "location": {
+                                "description": "The city and state, e.g. San Francisco, CA",
+                                "type": "string",
+                            },
+                            "unit": {
+                                "description": "Unit for the output - one of (celsius, fahrenheit)",
+                                "type": "string",
+                            },
+                        },
+                    },
+                },
+                {
+                    "description": "Get the current weather in a given location",
+                    "name": "x",
+                    "input_schema": {
+                        "type": "object",
+                        "properties": {
+                            "location": {
+                                "description": "The city and state, e.g. San Francisco, CA",
+                                "type": "string",
+                            },
+                            "unit": {
+                                "description": "Unit for the output - one of (celsius, fahrenheit)",
+                                "type": "string",
+                            },
+                        },
+                    },
+                },
+                {
+                    "description": "Get the current weather in a given location",
+                    "name": "x",
+                    "input_schema": {
+                        "type": "object",
+                        "properties": {
+                            "location": {
+                                "description": "The city and state, e.g. San Francisco, CA",
+                                "type": "string",
+                            },
+                            "unit": {
+                                "description": "Unit for the output - one of (celsius, fahrenheit)",
+                                "type": "string",
+                            },
+                        },
+                    },
+                },
+            ],
             top_k=5,
             top_p=0.7,
         )
@@ -199,6 +307,60 @@ class TestAsyncMessages:
             stream=False,
             system="Today's date is 2024-01-01.",
             temperature=1,
+            tool_choice={"type": "auto"},
+            tools=[
+                {
+                    "description": "Get the current weather in a given location",
+                    "name": "x",
+                    "input_schema": {
+                        "type": "object",
+                        "properties": {
+                            "location": {
+                                "description": "The city and state, e.g. San Francisco, CA",
+                                "type": "string",
+                            },
+                            "unit": {
+                                "description": "Unit for the output - one of (celsius, fahrenheit)",
+                                "type": "string",
+                            },
+                        },
+                    },
+                },
+                {
+                    "description": "Get the current weather in a given location",
+                    "name": "x",
+                    "input_schema": {
+                        "type": "object",
+                        "properties": {
+                            "location": {
+                                "description": "The city and state, e.g. San Francisco, CA",
+                                "type": "string",
+                            },
+                            "unit": {
+                                "description": "Unit for the output - one of (celsius, fahrenheit)",
+                                "type": "string",
+                            },
+                        },
+                    },
+                },
+                {
+                    "description": "Get the current weather in a given location",
+                    "name": "x",
+                    "input_schema": {
+                        "type": "object",
+                        "properties": {
+                            "location": {
+                                "description": "The city and state, e.g. San Francisco, CA",
+                                "type": "string",
+                            },
+                            "unit": {
+                                "description": "Unit for the output - one of (celsius, fahrenheit)",
+                                "type": "string",
+                            },
+                        },
+                    },
+                },
+            ],
             top_k=5,
             top_p=0.7,
         )
@@ -273,6 +435,60 @@ class TestAsyncMessages:
             stop_sequences=["string", "string", "string"],
             system="Today's date is 2024-01-01.",
             temperature=1,
+            tool_choice={"type": "auto"},
+            tools=[
+                {
+                    "description": "Get the current weather in a given location",
+                    "name": "x",
+                    "input_schema": {
+                        "type": "object",
+                        "properties": {
+                            "location": {
+                                "description": "The city and state, e.g. San Francisco, CA",
+                                "type": "string",
+                            },
+                            "unit": {
+                                "description": "Unit for the output - one of (celsius, fahrenheit)",
+                                "type": "string",
+                            },
+                        },
+                    },
+                },
+                {
+                    "description": "Get the current weather in a given location",
+                    "name": "x",
+                    "input_schema": {
+                        "type": "object",
+                        "properties": {
+                            "location": {
+                                "description": "The city and state, e.g. San Francisco, CA",
+                                "type": "string",
+                            },
+                            "unit": {
+                                "description": "Unit for the output - one of (celsius, fahrenheit)",
+                                "type": "string",
+                            },
+                        },
+                    },
+                },
+                {
+                    "description": "Get the current weather in a given location",
+                    "name": "x",
+                    "input_schema": {
+                        "type": "object",
+                        "properties": {
+                            "location": {
+                                "description": "The city and state, e.g. San Francisco, CA",
+                                "type": "string",
+                            },
+                            "unit": {
+                                "description": "Unit for the output - one of (celsius, fahrenheit)",
+                                "type": "string",
+                            },
+                        },
+                    },
+                },
+            ],
             top_k=5,
             top_p=0.7,
         )
